Autocleanup as view (#875)
* not needed imports removed

* cleanup view added

* ADMIN_ACCESS_TOKEN added to the default config file

* generateautocleanup management function added

* docker settings updated to be able to generate a script that triggers autocleanup

* bind updated

* bind removed for now

* bind removed for now

* not needed imports removed

* binding added

* binding files replaces with binding directories

* generateautocleanup updated to generate a new script also when there has been token update

* some code refinement done; todo comment added

* function name udpated

* email about validation expiration updated;

* sending email to admins in case of failed clean up action

* testing validation autoclenaup moved to api tests, because that's the proper location now

* autocleanupvalidations.py removed, as the command is now triggered by an api call and not as a management function

* script updated to return proper message when there is no user for which it can be generated

* testing generating autocleanup script added

* testing sending email about autocleanup failure added

* sending email about expiring validation only when there is an email

* help page updated - modifying multiple validaitons explained

* help page updated - modifying multiple validaitons explained

* needed pictures added
